♻️ Refactor: qustion 도메인 부족한 타이핑 마무리
